+++ Social-Trading/Clean-UI/Client/httpInterface.js
