+++ Social-Trading/Clean-UI/Client/webAppInterface.js
