+++ Social-Trading/Clean-UI/Client/webSocketsInterface.js
