+++ Social-Trading/Clean-UI/SocialTradingApp.js
