+++ Social-Trading/Clean-UI/UI/Globals.js
