+++ Social-Trading/Clean-UI/UI/WebApp.js
