+++ Social-Trading/Clean-UI/UI/WebDebugLog.js
