+++ Social-Trading/Clean-UI/UI/WebSocketsWebAppClient.js
